Move the groupedToolCallCallbacks to the persona instead of jupyterlite-ai
