+++ packages/ai/src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './clear-button';
 export * from './model-select';
+export * from './save-button';
 export * from './stop-button';
-export * from './usage-display';
 export * from './tool-select';
+export * from './usage-display';
