+++ packages/persona/package.json
@@ -67,6 +67,7 @@
     "@mui/icons-material": "^7",
     "@mui/material": "^7",
     "ai": "^7.0.59",
+    "jupyter-chat-components": "^0.6.0",
     "jupyter-mcp-manager": "^0.2.0",
     "jupyter-secrets-manager": "^0.5.0",
     "react": "^18.3.1"
